% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_f.R
\name{check_limma}
\alias{check_limma}
\title{Check for 'limma' availability}
\usage{
check_limma()
}
\description{
checks if the 'limma' package is installed. If not already installed,
limma will be installed automatically.
}
\details{
Check for package dependency
}
\seealso{
For more information, please refer to the package's documentation and the tutorial: \url{https://software.klausjung-lab.de/}.
}
\author{
Sergej Ruff
}
\keyword{internal}
