% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_data.R
\name{GA.diagplot}
\alias{GA.diagplot}
\title{Diagnostic plot for comparison of two covariance matrices}
\usage{
GA.diagplot(R, Rt, eps = 0.05, col.method = "traficlight", color = c(0,
  8), top = "")
}
\arguments{
\item{R}{Specified covariance matrix}

\item{Rt}{Covariance matrix of the data generated by the genetic
algorithm}

\item{eps}{Permitted difference between the entries of two
matrices. Must only be specified if col.method="traficlight"}

\item{col.method}{Method to use for color scaling the difference
between the matrices. If method="traficlight" only two colors
are used, indicating whether the entries deviated at least by a
difference of eps. If method="updown" a discrete gray scale is
used.}

\item{color}{Value of two color that are used if method="traficlight"}

\item{top}{Specifies the main title of the plot}
}
\description{
A diagnostic plot that compares the entries of two covariance
matrices using a color scale.
}
\details{
A diagnostic plot that compares the entries of two covariance
matrices using a color scale.
}
\examples{
### A simple example
X0 <- start.matrix(p = c(0.1, 0.2, 0.3), k = 100)
X <- iter.matrix(X0, R = diag(3))
GA.diagplot(diag(3), Rt = X$Rt)
}
\author{
Jochen Kruppa, Klaus Jung
}
