% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_data.R
\name{iter.matrix}
\alias{iter.matrix}
\title{Genetic algorithm for generating correlated binary data}
\usage{
iter.matrix(X0, V, tloc = 1000, e.break = 1e-04, plt = TRUE,
  perc = TRUE)
}
\arguments{
\item{X0}{Start matrix with specified marginal probabilities. Can
be generated by \code{\link{start.matrix}}.}

\item{V}{Desired covariance matrix the data should have after
running the genetic algorithm}

\item{tloc}{Maximum number of iterations after which the genetic
algorithm stops}

\item{e.break}{Minimum error between the covariance of the
iterated data matrix and V.}

\item{plt}{Boolean parameter that indicates whether to plot
e.break versus the iteration step}

\item{perc}{Boolean parameter that indicates whether to print the
percentage of iteration steps relativ to Tloc have been carried
out}
}
\value{
A list with four entries:
 \describe{
\item{\emph{X1}}{Final data with specified marginal probabilities and a covariance as close as possible to V} 
\item{\emph{t}}{Number of performed iteration steps (t <= Tloc)}
\item{\emph{V1}}{Covariance matrix of X1}
\item{\emph{e}}{Final RSME error between desired and achieved covariance matrix }
}
}
\description{
Starts the genetic algorithm based on a start matrix with
specified marginal probabilities.
}
\details{
In each step, the genetic algorithm swaps two randomly selected
entries in each column of X0. Thus it can be guaranteed that the
marginal probabilities do not change. If the covariance matrix of
X0(t) is gets to V in step t X0(t) replaces X0(t-1).
}
\examples{
### A simple example
X0 = start.matrix(n = 100, p = c(0.5, 0.6))
X = iter.matrix(X0, V = diag(2))
}
\author{
Jochen Kruppa, Klaus Jung
}
