% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_data.R
\name{rmvbinary_EP}
\alias{rmvbinary_EP}
\title{Title}
\usage{
rmvbinary_EP(n, R, p)
}
\arguments{
\item{n}{}

\item{R}{}

\item{p}{}
}
\description{
Description
}
\details{
Details
}
\examples{
## Here should be an example
}
\references{
Emrich, L.J., Piedmonte, M.R.: A method for generating
  highdimensional multivariate binary variates. The American
  Statistician 45(4), 302 (1991).
}
\author{
Jochen Kruppa, Klaus Jung
}
