% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_data.R
\name{rmvbinary_QA}
\alias{rmvbinary_QA}
\title{Simulating correlated binary variables using the algorithm by Qaqish (2003)}
\usage{
rmvbinary_QA(n, R, p)
}
\arguments{
\item{n}{Number of wanted samples to be drawn}

\item{R}{Correlation matrix}

\item{p}{Vector of marginal frequencies}
}
\value{
A (n x p)-Matrix with with entries 0 and 1 according to
  the specified marginal probabilities p.
}
\description{
Description
}
\details{
Details
}
\examples{
## Generation of a random sample
rmvbinary_QA(n = 10, R = diag(2), p = c(0.5, 0.6))
}
\references{
Qaqish, B. F.: A family of multivariate binary
  distributions for simulating correlated binary variables with
  specified marginal means and correlations, Biometrika 90 (2),
  455-463 (2003)
}
\author{
Jochen Kruppa, Klaus Jung
}
