% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{robust_covid_markers}
\alias{robust_covid_markers}
\title{Robust COVID-19 Markers Dataset}
\description{
Robust marker data from COVID-19 analysis using CLR transformation, 5 PCs, and 1000 features.
}
\keyword{internal}
