% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{robust_covid_markers_03trim}
\alias{robust_covid_markers_03trim}
\title{Robust COVID-19 Markers (03 Trim) Dataset}
\description{
Robust marker data with 0.3 trimming from COVID-19 analysis using CLR transformation.
}
\keyword{internal}
