% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{scTC_eff_clr}
\alias{scTC_eff_clr}
\title{COVID-19 CLR Transformation Marker Effects (Internal)}
\format{
A data frame with marker genes as rows and the following columns:
\describe{
  \item{CellAnnotation}{Additional cell annotation information}
  \item{X}{Row identifier}
  \item{avg_log2FC}{Average log2 fold-change}
  \item{gene}{Gene identifier}
  \item{p_val}{Raw p-value}
  \item{p_val_adj}{Adjusted p-value}
  \item{pct.1}{Percentage of cells expressing the gene in cluster}
  \item{pct.2}{Percentage of cells expressing the gene in other clusters}
  \item{cluster}{Cluster assignment}
}
}
\description{
Internal dataset containing marker gene effects using CLR transformation
(5 PCs, 1000 features) for evaluating trimming effects in scTrimClust.
Used by \code{\link{scTC_trim_effect}}.
}
\keyword{internal}
