% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{scTC_eff_log}
\alias{scTC_eff_log}
\title{COVID-19 LogNormalized Marker Effects (Internal)}
\format{
A data frame with the same structure as \code{scTC_eff_clr}
}
\description{
Internal dataset containing marker gene effects using LogNormalization
(5 PCs, 1000 features) for evaluating trimming effects in scTrimClust.
Used by \code{\link{scTC_trim_effect}}.
}
\keyword{internal}
