% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_data.R
\name{start_matrix}
\alias{start_matrix}
\title{Setup of the start matrix}
\usage{
start_matrix(p, k)
}
\arguments{
\item{p}{Marginal probabilities of the start matrix}

\item{k}{Number of rows to be generated}
}
\value{
A (n x p)-Matrix with with entries 0 and 1 according to
  the specified marginal probabilities p.
}
\description{
Generation of the start matrix with n rows and specified marginal
probabilities p.
}
\details{
The start matrix needs to be setup for further use in the genetic
algorithm implemented in the function \code{\link{iter_matrix}}. For
high-dimensional cases or if the marginal probabilities have
multiple decimal places, the number n of rows should be large (up
to multiple thousand).
}
\examples{
X0 <- start_matrix(p = c(0.5, 0.6), k = 100)
}
\author{
Jochen Kruppa, Klaus Jung
}
