% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_RHD.R
\name{summary_RHD}
\alias{summary_RHD}
\title{Summary of RHighDim function}
\usage{
summary_RHD(object, ...)
}
\arguments{
\item{object}{An object provided by the RHighDim function.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
No value
}
\description{
Summary of RHighDim function
}
\details{
Summarizes the test results obtained by the RHighDim function.
}
\examples{
### Global comparison of a set of 100 genes between two experimental groups.
X1 = matrix(rnorm(1000, 0, 1), 10, 100)
X2 = matrix(rnorm(1000, 0.1, 1), 10, 100)
RHD = RHighDim (X1, X2, paired=FALSE)
summary_RHD(RHD)
}
\references{
Brunner, E (2009) Repeated measures under non-sphericity. \emph{Proceedings of the 6th St. Petersburg Workshop on Simulation}, 605-609.

Jung K, Becker B, Brunner B and Beissbarth T (2011) Comparison of Global Tests for Functional Gene Sets in Two-Group Designs and Selection of Potentially Effect-causing Genes. \emph{Bioinformatics}, \strong{27}, 1377-1383. \doi{10.1093/bioinformatics/btr152}
}
\author{
Klaus Jung
}
