% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_limma.R
\name{check_limma}
\alias{check_limma}
\title{Check for 'limma' availability}
\usage{
check_limma()
}
\description{
checks if the 'limma' package is installed. If not already installed,
limma will be installed automatically.
}
\details{
Check for package dependency
}
\author{
Sergej Ruff
}
\keyword{internal}
