% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_data.R
\name{rmvbinary_EP}
\alias{rmvbinary_EP}
\title{Simulating correlated binary variables using the algorithm by Emrich and Piedmonte (1991)}
\usage{
rmvbinary_EP(n, R, p)
}
\arguments{
\item{n}{Number of wanted samples to be drawn}

\item{R}{Correlation matrix}

\item{p}{Vector of marginal frequencies}
}
\value{
A (n x p)-Matrix with with entries 0 and 1 according to
  the specified marginal probabilities p.
}
\description{
Description
}
\details{
Details
}
\examples{
## Generation of a random sample
rmvbinary_EP(n = 10, R = diag(2), p = c(0.5, 0.6))
}
\references{
Emrich, L.J., Piedmonte, M.R.: A method for generating
  highdimensional multivariate binary variates. The American
  Statistician 45(4), 302 (1991).
}
\author{
Jochen Kruppa, Klaus Jung
}
