% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_data.R
\name{rmvbinary_QA}
\alias{rmvbinary_QA}
\title{Title}
\usage{
rmvbinary_QA(n, R, p)
}
\arguments{
\item{n}{}

\item{R}{}

\item{p}{}
}
\description{
Description
}
\details{
Details
}
\examples{
## Here should be an example
}
\references{
Qaqish, B. F.: A family of multivariate binary
  distributions for simulating correlated binary variables with
  specified marginal means and correlations, Biometrika 90 (2),
  455-463 (2003)
}
\author{
Jochen Kruppa, Klaus Jung
}
