% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestStatSimple.R
\name{TestStatSimple}
\alias{TestStatSimple}
\title{Calculates test statistic for RHighDim}
\usage{
TestStatSimple(Y, H)
}
\arguments{
\item{Y}{A matrix.}

\item{H}{A matrix.}
}
\value{
A list.
}
\description{
Calculates test statistic for RHighDim
}
