% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestStatSP.R
\name{TestStatSP}
\alias{TestStatSP}
\title{Calculates test statistic for RHighDim}
\usage{
TestStatSP(Y1, Y2)
}
\arguments{
\item{Y1}{A matrix.}

\item{Y2}{A matrix.}
}
\value{
A list.
}
\description{
Calculates test statistic for RHighDim
}
